Update README.md
🎯 Takeaways
This project teach me the first harmonic of data cleaning, explorative analysis, and visualization. It strengthened my analytical science and take fire my heat for unveil stories within data.  Pizza Sales is not just a task ; it ’s a milepost that position the foundation for to a greater extent sophisticated endeavor in my analytics journey.